tools: /dev/rpmsg-omx demo app
This is a user space application that demonstrates usage
of the /dev/rpmsg-omx device.

The app creates a remote OMX instance, and once connection
is established, it starts ping-ponging raw messages with that new OMX
instance (note: currently a generic "OMX" name is used while connecting.
this can be changed to a specific OMX instance name, such as
"h264_decode", when we implement the entire GetHandle function within
the context of the connect ioctl).

We don't really invoke meaningful OMX remote API yet, but this
should be enough infrastructure to start doing real OMX work on both sides
(A9 and Ducati).

Btw I'm using eventfd(2), which I'm not sure bionic supports. Anyway it
can be trivially replaced by pipe(2).

Change-Id: Ib01a0efff6be26108422fdfc1ff5fe3afb3eb67e
Signed-off-by: Ohad Ben-Cohen <ohad@wizery.com>